civic_callbot/requirements.txt
flask
twilio
sentence-transformers
faiss-cpu
openai
requests
